# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Todo application REST-API",
        "title": "todoky",
        "contact": {},
        "version": "1.0"
    },
    "host": "172.21.253.37:5050",
    "basePath": "/api/v1",
    "paths": {
...
        "/tasks": {
            "get": {
                "description": "Получение задач с опциональной пагинацией",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Получения списка задач",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Размер страницы с задачами",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Смещение страницы с задачами",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное получение списка задач",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_features_tasks_transport_http.TaskDTOResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_KyoshiBlame_TodoKy_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_KyoshiBlame_TodoKy_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Создать новую задачу в системе",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Создать задачу",
                "parameters": [
                    {
                        "description": "CreateTask тело запроса",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_features_tasks_transport_http.CreateTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Успешно созданная задача",
                        "schema": {
                            "$ref": "#/definitions/internal_features_tasks_transport_http.CreateTaskResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_KyoshiBlame_TodoKy_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Author not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_KyoshiBlame_TodoKy_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal sercer error",
                        "schema": {
                            "$ref": "#/definitions/github_com_KyoshiBlame_TodoKy_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/tasks/{id}": {
            "get": {
                "description": "Получить задачу из системы",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Получить задачу",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID задачи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное получение задачи",
                        "schema": {
                            "$ref": "#/definitions/internal_features_tasks_transport_http.CreateTaskResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_KyoshiBlame_TodoKy_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal sercer error",
                        "schema": {
                            "$ref": "#/definitions/github_com_KyoshiBlame_TodoKy_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удалить существующую задачу в системе",
                "tags": [
                    "Tasks"
                ],
                "summary": "Удалить задачу",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID задачи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Успешно удалена задача"
                    },
                    "404": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_KyoshiBlame_TodoKy_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal sercer error",
                        "schema": {
                            "$ref": "#/definitions/github_com_KyoshiBlame_TodoKy_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Изменить существующую задачу в системе\n### Логика обновления полей (Three-state logic):\n1. **Поле не передано**: `description` игнорируется, значение в БД не меняется\n2. **Явно передано значение**: `\"description\":\"помыть лапки собаке\"` - устанавливает новый номер в БД\n3. **Передано NULL**: `\"description\":null` - очистит поле в БД (set NULL)\nОграниечение: `title` не может быть выставлен как NULL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Изменнение задачи",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID изменяемой задачи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "PatchTaskRequest тело запроса",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_features_tasks_transport_http.PatchTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное изменение задачи",
                        "schema": {
                            "$ref": "#/definitions/internal_features_tasks_transport_http.TaskPatchedResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_KyoshiBlame_TodoKy_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_KyoshiBlame_TodoKy_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_KyoshiBlame_TodoKy_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_KyoshiBlame_TodoKy_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_KyoshiBlame_TodoKy_internal_core_transport_http_response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "full error text"
                },
                "message": {
                    "type": "string",
                    "example": "short text for human"
                }
            }
        },
...
        "internal_features_tasks_transport_http.CreateTaskRequest": {
            "type": "object",
            "required": [
                "author_user_id",
                "title"
            ],
            "properties": {
                "author_user_id": {
                    "type": "integer",
                    "example": 1
                },
                "description": {
                    "type": "string",
                    "maxLength": 1000,
                    "minLength": 1,
                    "example": "Купить рис, курицу и молока"
                },
                "title": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1,
                    "example": "Сходить в магазин"
                }
            }
        },
        "internal_features_tasks_transport_http.CreateTaskResponse": {
            "type": "object",
            "properties": {
                "author_user_id": {
                    "type": "integer",
                    "example": 3
                },
                "completed": {
                    "type": "boolean",
                    "example": false
                },
                "completed_at": {
                    "type": "string",
                    "example": ""
                },
                "created_at": {
                    "type": "string",
                    "example": "04.08.2026 17:01:00"
                },
                "description": {
                    "type": "string",
                    "example": "Купить рис, курицу и молоко"
                },
                "id": {
                    "type": "integer",
                    "example": 3
                },
                "title": {
                    "type": "string",
                    "example": "Сходить в магазин"
                },
                "version": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "internal_features_tasks_transport_http.PatchTaskRequest": {
            "type": "object",
            "properties": {
                "completed": {
                    "type": "boolean"
                },
                "description": {
                    "type": "string",
                    "example": "Погулять с пушком в новом парке"
                },
                "title": {
                    "type": "string",
                    "example": "Сходить в магазин"
                }
            }
        },
        "internal_features_tasks_transport_http.TaskDTOResponse": {
            "type": "object",
            "properties": {
                "author_user_id": {
                    "type": "integer",
                    "example": 3
                },
                "completed": {
                    "type": "boolean",
                    "example": false
                },
                "completed_at": {
                    "type": "string",
                    "example": ""
                },
                "created_at": {
                    "type": "string",
                    "example": "04.08.2026 17:01:00"
                },
                "description": {
                    "type": "string",
                    "example": "Купить рис, курицу и молоко"
                },
                "id": {
                    "type": "integer",
                    "example": 3
                },
                "title": {
                    "type": "string",
                    "example": "Сходить в магазин"
                },
                "version": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "internal_features_tasks_transport_http.TaskPatchedResponse": {
            "type": "object",
            "properties": {
                "author_user_id": {
                    "type": "integer",
                    "example": 3
                },
                "completed": {
                    "type": "boolean",
                    "example": false
                },
                "completed_at": {
                    "type": "string",
                    "example": ""
                },
                "created_at": {
                    "type": "string",
                    "example": "04.08.2026 17:01:00"
                },
                "description": {
                    "type": "string",
                    "example": "Купить рис, курицу и молоко"
                },
                "id": {
                    "type": "integer",
                    "example": 3
                },
                "title": {
                    "type": "string",
                    "example": "Сходить в магазин"
                },
                "version": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
    }
}